fix warning when building on 32-bit

+++ gtk/gtkclipboard.c
@@ -1780,7 +1780,7 @@ gtk_clipboard_wait_for_targets (GtkClipboard  *clipboard,
  	{
  	  clipboard->n_cached_targets = tmp_n_targets;
  	  clipboard->cached_targets = g_memdup2 (tmp_targets,
- 						(guint)tmp_n_targets * sizeof (GdkAtom));
+ 						(long)tmp_n_targets * sizeof (GdkAtom));
  	}
  
       if (n_targets)
